+++ lib/di/src/main/scala/org/bitbucket/pshirshov/izumi/di/DIContext.scala
@@ -1,0 +1,9 @@
+package org.bitbucket.pshirshov.izumi.di
+
+import org.bitbucket.pshirshov.izumi.di.model.DIKey
+
+trait DIContext {
+  def instance[T: Tag](key: DIKey): T
+
+  def subInjector(): Injector
+}
